+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2021-05-07
 ### Added
 - Encode base32 without padding.
 - Encode base32hex without padding.